+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- networks: Sort interfaces by port names and disregard order in uci config
+
 ## [5.0.0] - 2022-05-18
 ### Added
 - lan: allow managing of DHCPv4 static leases (create, update, delete)